refactor(#16): 스웨거 인터페이스 분리
