+++ src/main/java/org/example/stellog/member/api/MemberController.java
@@ -1,13 +1,8 @@
 package org.example.stellog.member.api;
 
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
+import org.example.stellog.gcs.application.GCSService;
 import org.example.stellog.global.annotation.AuthenticatedEmail;
-import org.example.stellog.global.gcs.GCSService;
 import org.example.stellog.global.template.RspTemplate;
 import org.example.stellog.member.api.dto.request.MemberUpdateReqDto;
 import org.example.stellog.member.api.dto.response.MemberInfoResDto;
@@ -22,15 +17,10 @@ import java.io.IOException;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/v1/members")
-@Tag(name = "Member", description = "회원 관련 API")
-public class MemberController {
+public class MemberController implements MemberControllerDocs {
     private final MemberService memberService;
     private final GCSService gcsService;
 
-    @Operation(
-            summary = "회원 닉네임 수정",
-            description = "회원 닉네임을 수정합니다"
-    )
     @PutMapping("/nickname")
     public RspTemplate<String> updateMemberNickName(@AuthenticatedEmail String email, @RequestBody MemberUpdateReqDto memberUpdateReqDto) {
         memberService.updateMemberNickNme(email, memberUpdateReqDto);
@@ -40,10 +30,6 @@ public class MemberController {
         );
     }
 
-    @Operation(
-            summary = "회원 프로필 이미지 수정",
-            description = "회원 프로필 이미지를 수정합니다."
-    )
     @PutMapping("/profile")
     public RspTemplate<String> updateProfileImg(@AuthenticatedEmail String email,
                                                 @RequestParam("file") MultipartFile file) throws IOException {
@@ -56,16 +42,6 @@ public class MemberController {
         );
     }
 
-    @Operation(
-            summary = "회원 정보 조회",
-            description = "회원 정보를 조회합니다."
-    )
-    @ApiResponse(
-            content = @Content(
-                    mediaType = "application/json",
-                    schema = @Schema(implementation = MemberInfoResDto.class)
-            )
-    )
     @GetMapping("/info")
     public RspTemplate<MemberInfoResDto> getMember(@AuthenticatedEmail String email) {
         return new RspTemplate<>(
@@ -75,10 +51,6 @@ public class MemberController {
         );
     }
 
-    @Operation(
-            summary = "회원 목록 조회",
-            description = "회원 목록을 조회합니다. 방 생성 시 참여할 회원을 선택하기 위해 사용됩니다. 이름 중 하나 글자만 입력해도 조회가 됩니다."
-    )
     @GetMapping("/list")
     public RspTemplate<MemberListResDto> getMemberList(@AuthenticatedEmail String email,
                                                        @RequestParam("name") String name) {
