+++ src/main/java/org/example/stellog/review/api/ReviewController.java
@@ -1,7 +1,5 @@
 package org.example.stellog.review.api;
 
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.example.stellog.global.annotation.AuthenticatedEmail;
 import org.example.stellog.global.template.RspTemplate;
@@ -14,15 +12,10 @@ import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequiredArgsConstructor
-@Tag(name = "Review", description = "리뷰 관련 API")
 @RequestMapping("/api/v1/reviews")
-public class ReviewController {
+public class ReviewController implements ReviewControllerDocs {
     private final ReviewService reviewService;
 
-    @Operation(
-            summary = "리뷰 생성",
-            description = "리뷰를 생성합니다."
-    )
     @PostMapping("{roomId}")
     public RspTemplate<Void> createReview(@AuthenticatedEmail String email, @PathVariable Long roomId, @RequestBody ReviewReqDto reviewReqDto) {
         reviewService.createReview(email, roomId, reviewReqDto);
@@ -32,10 +25,6 @@ public class ReviewController {
                 "리뷰가 성공적으로 생성되었습니다.");
     }
 
-    @Operation(
-            summary = "스타벅스 별 리뷰 목록 조회",
-            description = "스타벅스 별 리뷰 목록을 조회합니다."
-    )
     @GetMapping("/starbucks/{starbucksId}")
     public RspTemplate<ReviewListResDto> getAllReviewsByStarbucksId(@AuthenticatedEmail String email, @PathVariable Long starbucksId) {
         return new RspTemplate<>(
@@ -44,10 +33,6 @@ public class ReviewController {
                 reviewService.getAllReviewsByStarbucksId(email, starbucksId));
     }
 
-    @Operation(
-            summary = "방 별 리뷰 목록 조회",
-            description = "방 별 리뷰 목록을 조회합니다."
-    )
     @GetMapping("/room/{roomId}")
     public RspTemplate<ReviewListResDto> getAllReviewsByRoomId(@AuthenticatedEmail String email, @PathVariable Long roomId) {
         return new RspTemplate<>(
@@ -56,10 +41,6 @@ public class ReviewController {
                 reviewService.getAllReviewsByRoomId(email, roomId));
     }
 
-    @Operation(
-            summary = "방과 스타벅스 별 리뷰 목록 조회",
-            description = "방과 스타벅스 별 리뷰 목록을 조회합니다."
-    )
     @GetMapping("{roomId}/{starbucksId}")
     public RspTemplate<ReviewListResDto> getReviewsByRoomIdAndStarbucksId(@AuthenticatedEmail String email, @PathVariable Long roomId, @PathVariable Long starbucksId) {
         return new RspTemplate<>(
@@ -68,10 +49,6 @@ public class ReviewController {
                 reviewService.getReviewsByRoomIdAndStarbucksId(email, roomId, starbucksId));
     }
 
-    @Operation(
-            summary = "리뷰 상세 조회",
-            description = "리뷰를 상세 조회합니다."
-    )
     @GetMapping("/{reviewId}")
     public RspTemplate<ReviewInfoResDto> getReviewById(@AuthenticatedEmail String email, @PathVariable Long reviewId) {
         return new RspTemplate<>(
@@ -80,10 +57,6 @@ public class ReviewController {
                 reviewService.getReviewById(email, reviewId));
     }
 
-    @Operation(
-            summary = "리뷰 수정",
-            description = "리뷰를 수정합니다. 리뷰를 작성한 사용자만 수정할 수 있습니다."
-    )
     @PutMapping({"/{reviewId}"})
     public RspTemplate<Void> updateReview(@AuthenticatedEmail String email, @PathVariable Long reviewId, @RequestBody ReviewReqDto reviewReqDto) {
         reviewService.updateReview(email, reviewId, reviewReqDto);
@@ -92,10 +65,6 @@ public class ReviewController {
                 "리뷰가 성공적으로 수정되었습니다.");
     }
 
-    @Operation(
-            summary = "리뷰 삭제",
-            description = "리뷰를 삭제합니다. 리뷰를 작성한 사용자만 삭제할 수 있습니다."
-    )
     @DeleteMapping("/{reviewId}")
     public RspTemplate<Void> deleteReview(@AuthenticatedEmail String email, @PathVariable Long reviewId) {
         reviewService.deleteReview(email, reviewId);
@@ -104,10 +73,6 @@ public class ReviewController {
                 "리뷰가 성공적으로 삭제되었습니다.");
     }
 
-    @Operation(
-            summary = "리뷰 좋아요",
-            description = "리뷰에 좋아요를 누릅니다."
-    )
     @PostMapping("/like/{reviewId}")
     public RspTemplate<Void> likeReview(@AuthenticatedEmail String email, @PathVariable Long reviewId) {
         reviewService.likeReview(email, reviewId);
@@ -116,10 +81,6 @@ public class ReviewController {
                 "리뷰에 좋아요를 눌렀습니다.");
     }
 
-    @Operation(
-            summary = "리뷰 좋아요 취소",
-            description = "리뷰에 좋아요를 취소합니다."
-    )
     @DeleteMapping("/like/{reviewId}")
     public RspTemplate<Void> unlikeReview(@AuthenticatedEmail String email, @PathVariable Long reviewId) {
         reviewService.unlikeReview(email, reviewId);
