+++ src/main/java/org/example/stellog/room/api/RoomController.java
@@ -1,10 +1,5 @@
 package org.example.stellog.room.api;
 
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.example.stellog.global.annotation.AuthenticatedEmail;
 import org.example.stellog.global.template.RspTemplate;
@@ -17,15 +12,10 @@ import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequiredArgsConstructor
-@Tag(name = "Room", description = "방 관련 API")
 @RequestMapping("/api/v1/rooms")
-public class RoomController {
+public class RoomController implements RoomControllerDocs {
     private final RoomService roomService;
 
-    @Operation(
-            summary = "방 개설",
-            description = "공개 범위를 설정하여 방을 개설합니다."
-    )
     @PostMapping
     public RspTemplate<Void> createRoom(@AuthenticatedEmail String email, @RequestBody RoomReqDto roomReqDto) {
         roomService.createRoom(email, roomReqDto);
@@ -34,10 +24,6 @@ public class RoomController {
                 "방이 성공적으로 생성되었습니다.");
     }
 
-    @Operation(
-            summary = "사용자가 생성한 방 목록 조회",
-            description = "현재 로그인 한 사용자의 모든 방 목록을 조회합니다."
-    )
     @GetMapping
     public RspTemplate<RoomListResDto> getAllRoomByEmail(@AuthenticatedEmail String email) {
         return new RspTemplate<>(
@@ -46,16 +32,6 @@ public class RoomController {
                 roomService.getAllRoomByEmail(email));
     }
 
-    @Operation(
-            summary = "방 상세 정보 조회",
-            description = "사용자의 방 상세 정보를 조회 합니다."
-    )
-    @ApiResponse(
-            content = @Content(
-                    mediaType = "application/json",
-                    schema = @Schema(implementation = RoomDetailResDto.class)
-            )
-    )
     @GetMapping("/detail/{roomId}")
     public RspTemplate<RoomDetailResDto> getRoomDetails(@AuthenticatedEmail String email, @PathVariable Long roomId) {
         return new RspTemplate<>(
@@ -64,10 +40,6 @@ public class RoomController {
                 roomService.getRoomDetails(email, roomId));
     }
 
-    @Operation(
-            summary = "방 수정",
-            description = "방을 수정합니다."
-    )
     @PutMapping("/{roomId}")
     public RspTemplate<Void> updateRoom(@AuthenticatedEmail String email, @PathVariable Long roomId, @RequestBody RoomReqDto roomReqDto) {
         roomService.updateRoom(email, roomId, roomReqDto);
@@ -76,10 +48,6 @@ public class RoomController {
                 "방을 성공적으로 수정하였습니다.");
     }
 
-    @Operation(
-            summary = "방 삭제",
-            description = "방을 삭제합니다."
-    )
     @DeleteMapping("/{roomId}")
     public RspTemplate<Void> deleteRoom(@AuthenticatedEmail String email, @PathVariable Long roomId) {
         roomService.deleteRoom(email, roomId);
